new code

+++ hal/include/ultrasonic.h
@@ -2,9 +2,11 @@
 #ifndef _ULTRASONIC_H_
 #define _ULTRASONIC_H_
 
+#include <stdint.h>
+
 void initializeUltrasonic();
 void* ultrasonicLoop();
-double getDistance();
+intmax_t getDistance();
 void ultrasonicShutdown();
 
 #endif
